+++ src/run-test.ts
@@ -1,7 +1,7 @@
 import { runTest } from './cypress/e2e/visual-regression-tests.cy';
-import { RunTest, TestProps } from './types';
+import { RunTest, TestConfig } from './types';
 
-export const runVisreg: RunTest = (props: TestProps) => {
+export const runVisreg: RunTest = (props: TestConfig) => {
     runTest(props);
 };